requirements.txt
streamlit>=1.28.0
snowflake-connector-python==3.6.0
pandas>=1.5.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=10.0.0
